requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.0.0
firebase-admin==6.6.0
python-dotenv==1.0.1
